Crate rustc_const_eval

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

constant evaluation on the HIR and code to validate patterns/matches

Note

This API is completely unstable and subject to change.

Modules

check_match 	 [
Experimental
]

pattern 	 [
Experimental
]

Structs

ConstContext 	 [
Experimental
]

Constants

DIAGNOSTICS 	 [
Experimental
]

Functions

compare_const_vals 	 [
Experimental
]

lookup_const_by_id 	 [
Experimental
]

DefId is the id of the constant.

Substs is the monomorphized substitutions for the expression.

provide 	 [
Experimental
]