Module rustc_const_eval::const_eval

Modules

error ğŸ”’

eval_queries ğŸ”’

fn_queries ğŸ”’

machine ğŸ”’

valtrees ğŸ”’

Structs

CompileTimeInterpreter

Extra machine state for CTFE, and the Machine instance

ConstEvalErr

When const-evaluation errors, this type is constructed with the resulting information, and then used to emit the error as a lint or hard error.

Enums

ConstEvalErrKind

The CTFE machine has some custom error kinds.

MemoryKind

ValTreeCreationError ğŸ”’

Constants

VALTREE_MAX_NODES ğŸ”’

Functions

const_caller_location ğŸ”’

deref_mir_constant ğŸ”’

eval_to_allocation_raw_provider

eval_to_const_value_raw_provider

eval_to_valtree ğŸ”’

Evaluates a constant and turns it into a type-level constant value.

is_parent_const_impl_raw

is_unstable_const_fn

Whether the def_id is an unstable const fn and what feature gate is necessary to enable it

provide

try_destructure_mir_constant ğŸ”’

Type Definitions

ValTreeCreationResult ğŸ”’