rustc_const_eval::const_eval::error

Function report

pub(super) fn report<'tcx, C, F>(
    ecx: &InterpCx<'tcx, CompileTimeMachine<'tcx>>,
    error: InterpErrorKind<'tcx>,
    span: Span,
    get_span_and_frames: C,
    mk: F,
) -> ErrorHandled
where
    C: FnOnce() -> (Span, Vec<FrameNote>),
    F: FnOnce(&mut Diag<'_>, Span, Vec<FrameNote>),

Create a diagnostic for a const eval error.

This will use the mk function for adding more information to the error. You can use it to add a stacktrace of current execution according to get_span_and_frames or just give context on where the const eval error happened.