Module rustc_const_eval::const_eval::eval_queries

Constants

NOTE_ON_UNDEFINED_BEHAVIOR_ERROR 🔒

Functions

eval_body_using_ecx 🔒

eval_to_allocation_raw_provider

eval_to_const_value_raw_provider

mk_eval_cx 🔒

The InterpCx is only meant to be used to do field and index projections into constants for simd_shuffle and const patterns in match arms. It never performs alignment checks.

op_to_const 🔒

This function converts an interpreter value into a constant that is meant for use in the type system.

turn_into_const_value 🔒