Struct rustc_const_eval::const_eval::machine::CompileTimeInterpreter

pub struct CompileTimeInterpreter<'mir, 'tcx> {
    pub(super) steps_remaining: usize,
    pub(super) stack: Vec<Frame<'mir, 'tcx, AllocId, ()>>,
    pub(super) can_access_statics: bool,
    pub(super) check_alignment: bool,
}

Extra machine state for CTFE, and the Machine instance

Fields
steps_remaining: usize

For now, the number of terminators that can be evaluated before we throw a resource exhaustion error.

Setting this to 0 disables the limit and allows the interpreter to run forever.

stack: Vec<Frame<'mir, 'tcx, AllocId, ()>>

The virtual call stack.

can_access_statics: bool

We need to make sure consts never point to anything mutable, even recursively. That is relied on for pattern matching on consts with references. To achieve this, two pieces have to work together:

Interning makes everything outside of statics immutable.

Pointers to allocations inside of statics can never leak outside, to a non-static global. This boolean here controls the second part.

check_alignment: bool

Whether to check alignment during evaluation.

Implementations

impl<'mir, 'tcx> CompileTimeInterpreter<'mir, 'tcx>

pub(crate) fn new(
const_eval_limit: Limit,
can_access_statics: bool,
check_alignment: bool
) -> Self