rustc_const_eval::const_eval::machine

Struct CompileTimeMachine

pub struct CompileTimeMachine<'tcx> {
    pub(super) num_evaluated_steps: usize,
    pub(super) stack: Vec<Frame<'tcx>>,
    pub(super) can_access_mut_global: CanAccessMutGlobal,
    pub(super) check_alignment: CheckAlignment,
    pub(crate) static_root_ids: Option<(AllocId, LocalDefId)>,
    union_data_ranges: FxHashMap<Ty<'tcx>, RangeSet>,
}

Extra machine state for CTFE, and the Machine instance.

Fields
num_evaluated_steps: usize

The number of terminators that have been evaluated.

This is used to produce lints informing the user that the compiler is not stuck. Set to usize::MAX to never report anything.

stack: Vec<Frame<'tcx>>

The virtual call stack.

can_access_mut_global: CanAccessMutGlobal

Pattern matching on consts with references would be unsound if those references could point to anything mutable. Therefore, when evaluating consts and when constructing valtrees, we ensure that only immutable global memory can be accessed.

check_alignment: CheckAlignment

Whether to check alignment during evaluation.

static_root_ids: Option<(AllocId, LocalDefId)>

If Some, we are evaluating the initializer of the static with the given LocalDefId, storing the result in the given AllocId. Used to prevent accesses to a static’s base allocation, as that may allow for self-initialization loops.

union_data_ranges: FxHashMap<Ty<'tcx>, RangeSet>

A cache of “data range” computations for unions (i.e., the offsets of non-padding bytes).

Implementations

impl<'tcx> CompileTimeMachine<'tcx>

pub(crate) fn new( can_access_mut_global: CanAccessMutGlobal, check_alignment: CheckAlignment, ) -> Self

impl<'tcx> CompileTimeMachine<'tcx>

pub fn best_lint_scope(&self, tcx: TyCtxt<'tcx>) -> HirId

Find the first stack frame that is within the current crate, if any. Otherwise, return the crate’s HirId