Trait rustc_const_eval::const_eval::InterpretationResult

pub trait InterpretationResult<'tcx> {
    // Required method
    fn make_result(
        mplace: MPlaceTy<'tcx>,
        ecx: &mut InterpCx<'tcx, CompileTimeMachine<'tcx>>,
    ) -> Self;
}

Required Methods

fn make_result( mplace: MPlaceTy<'tcx>, ecx: &mut InterpCx<'tcx, CompileTimeMachine<'tcx>>, ) -> Self

This function takes the place where the result of the evaluation is stored and prepares it for returning it in the appropriate format needed by the specific evaluation query.

Object Safety

This trait is not object safe.