Module rustc_const_eval::interpret::eval_context

Structs

Frame

A stack frame.

FrameInfo

What we store about a frame in an interpreter backtrace.

InterpCx

LocalState

State of a local variable including a memoized layout

SpanGuard 🔒

Enums

LocalValue

Current value of a local variable

StackPopCleanup

StackPopUnwind

Unwind information.

Functions

from_known_layout 🔒

Use the already known layout if given (but sanity check in debug mode), or compute the layout.

mir_assign_valid_types 🔒

Test if it is valid for a MIR assignment to assign src-typed place to dest-typed value. This test should be symmetric, as it is primarily about layout compatibility.