Struct rustc_const_eval::interpret::intern::InternVisitor

struct InternVisitor<'rt, 'mir, 'tcx, M: CompileTimeMachine<'mir, 'tcx, MemoryKind>> {
    ecx: &'rt mut InterpCx<'mir, 'tcx, M>,
    ref_tracking: &'rt mut RefTracking<(MPlaceTy<'tcx>, InternMode)>,
    leftover_allocations: &'rt mut FxIndexSet<AllocId>,
    mode: InternMode,
    inside_unsafe_cell: bool,
}

Fields
ecx: &'rt mut InterpCx<'mir, 'tcx, M>

The ectx from which we intern.

ref_tracking: &'rt mut RefTracking<(MPlaceTy<'tcx>, InternMode)>

Previously encountered safe references.

leftover_allocations: &'rt mut FxIndexSet<AllocId>

A list of all encountered allocations. After type-based interning, we traverse this list to also intern allocations that are only referenced by a raw pointer or inside a union.

mode: InternMode

The root kind of the value that we’re looking at. This field is never mutated for a particular allocation. It is primarily used to make as many allocations as possible read-only so LLVM can place them in const memory.

inside_unsafe_cell: bool

This field stores whether we are currently inside an UnsafeCell. This can affect the intern mode of references we encounter.

Implementations

impl<'rt, 'mir, 'tcx, M: CompileTimeMachine<'mir, 'tcx, MemoryKind>> InternVisitor<'rt, 'mir, 'tcx, M>

fn intern_shallow( &mut self, alloc_id: AllocId, mode: InternMode, ty: Option<Ty<'tcx>> ) -> Option<IsStaticOrFn>