rustc_const_eval::interpret::intern
Trait HasStaticRootDefId

pub trait HasStaticRootDefId {
    // Required method
    fn static_def_id(&self) -> Option<LocalDefId>;
}

Required Methods

fn static_def_id(&self) -> Option<LocalDefId>

Returns the DefId of the static item that is currently being evaluated. Used for interning to be able to handle nested allocations.