Function rustc_const_eval::interpret::intrinsics::eval_nullary_intrinsic

pub(crate) fn eval_nullary_intrinsic<'tcx>(
    tcx: TyCtxt<'tcx>,
    param_env: ParamEnv<'tcx>,
    def_id: DefId,
    substs: SubstsRef<'tcx>
) -> InterpResult<'tcx, ConstValue<'tcx>>

The logic for all nullary intrinsics is implemented here. These intrinsics don’t get evaluated inside an InterpCx and instead have their value computed directly from rustc internal info.