rustc_const_eval::interpret
Module place

Computations on places – field projections, going from mir::Place, and writing into a place. All high-level functions to write to memory work on places as destinations.

Modules

size_asserts 🔒

Structs

MPlaceTy

A MemPlace with its layout. Constructing it is only possible in this module.

MemPlace 🔒

PlaceTy

An evaluated place, together with its type.

Enums

MemPlaceMeta

Information required for the sound usage of a MemPlace.

Place 🔒

Traits

Writeable

The Weiteable trait describes interpreter values that can be written to.