Struct rustc_const_eval::interpret::AllocRef

pub struct AllocRef<'a, 'tcx, Prov: Provenance, Extra, Bytes: AllocBytes = Box<[u8]>> {
    alloc: &'a Allocation<Prov, Extra, Bytes>,
    range: AllocRange,
    tcx: TyCtxt<'tcx>,
    alloc_id: AllocId,
}

A reference to some allocation that was already bounds-checked for the given region and had the on-access machine hooks run.

Fields
alloc: &'a Allocation<Prov, Extra, Bytes>range: AllocRangetcx: TyCtxt<'tcx>alloc_id: AllocId
Implementations

impl<'tcx, 'a, Prov: Provenance, Extra, Bytes: AllocBytes> AllocRef<'a, 'tcx, Prov, Extra, Bytes>

pub fn read_scalar( &self, range: AllocRange, read_provenance: bool ) -> InterpResult<'tcx, Scalar<Prov>>

range is relative to this allocation reference, not the base of the allocation.

pub fn read_integer( &self, range: AllocRange ) -> InterpResult<'tcx, Scalar<Prov>>

range is relative to this allocation reference, not the base of the allocation.

pub fn read_pointer(&self, offset: Size) -> InterpResult<'tcx, Scalar<Prov>>

offset is relative to this allocation reference, not the base of the allocation.

pub fn get_bytes_strip_provenance<'b>(&'b self) -> InterpResult<'tcx, &'a [u8]>

range is relative to this allocation reference, not the base of the allocation.

pub(crate) fn has_provenance(&self) -> bool

Returns whether the allocation has provenance anywhere in the range of the AllocRef.