Struct rustc_const_eval::interpret::MPlaceTy

pub struct MPlaceTy<'tcx, Prov: Provenance = AllocId> {
    mplace: MemPlace<Prov>,
    pub layout: TyAndLayout<'tcx>,
    pub align: Align,
}

A MemPlace with its layout. Constructing it is only possible in this module.

Fields
mplace: MemPlace<Prov>layout: TyAndLayout<'tcx>align: Align

rustc does not have a proper way to represent the type of a field of a repr(packed) struct: it needs to have a different alignment than the field type would usually have. So we represent this here with a separate field that “overwrites” layout.align. This means layout.align should never be used for a MPlaceTy!

Implementations

impl<'tcx, Prov: Provenance> MPlaceTy<'tcx, Prov>

pub fn fake_alloc_zst(layout: TyAndLayout<'tcx>) -> Self

Produces a MemPlace that works for ZST but nothing else. Conceptually this is a new allocation, but it doesn’t actually create an allocation so you don’t need to worry about memory leaks.

pub fn offset_with_meta(
&self,
offset: Size,
meta: MemPlaceMeta<Prov>,
layout: TyAndLayout<'tcx>,
cx: &impl HasDataLayout
) -> InterpResult<'tcx, Self>

pub fn offset(
&self,
offset: Size,
layout: TyAndLayout<'tcx>,
cx: &impl HasDataLayout
) -> InterpResult<'tcx, Self>

pub fn from_aligned_ptr(
ptr: Pointer<Option<Prov>>,
layout: TyAndLayout<'tcx>
) -> Self

pub fn from_aligned_ptr_with_meta(
ptr: Pointer<Option<Prov>>,
layout: TyAndLayout<'tcx>,
meta: MemPlaceMeta<Prov>
) -> Self

pub(crate) fn len(&self, cx: &impl HasDataLayout) -> InterpResult<'tcx, u64>

pub(super) fn vtable(&self) -> Scalar<Prov>