Struct rustc_const_eval::interpret::validity::ValidityVisitor

struct ValidityVisitor<'rt, 'mir, 'tcx, M: Machine<'mir, 'tcx>> {
    path: Vec<PathElem>,
    ref_tracking: Option<&'rt mut RefTracking<MPlaceTy<'tcx, M::Provenance>, Vec<PathElem>>>,
    ctfe_mode: Option<CtfeValidationMode>,
    ecx: &'rt InterpCx<'mir, 'tcx, M>,
}

Fields
path: Vec<PathElem>

The path may be pushed to, but the part that is present when a function starts must not be changed! visit_fields and visit_array rely on this stack discipline.

ref_tracking: Option<&'rt mut RefTracking<MPlaceTy<'tcx, M::Provenance>, Vec<PathElem>>>ctfe_mode: Option<CtfeValidationMode>

None indicates this is not validating for CTFE (but for runtime).

ecx: &'rt InterpCx<'mir, 'tcx, M>
Implementations

impl<'rt, 'mir, 'tcx: 'mir, M: Machine<'mir, 'tcx>> ValidityVisitor<'rt, 'mir, 'tcx, M>

fn aggregate_field_path_elem( &mut self, layout: TyAndLayout<'tcx>, field: usize ) -> PathElem

fn with_elem<R>( &mut self, elem: PathElem, f: impl FnOnce(&mut Self) -> InterpResult<'tcx, R> ) -> InterpResult<'tcx, R>

fn read_immediate( &self, op: &OpTy<'tcx, M::Provenance>, expected: ExpectedKind ) -> InterpResult<'tcx, ImmTy<'tcx, M::Provenance>>

fn read_scalar( &self, op: &OpTy<'tcx, M::Provenance>, expected: ExpectedKind ) -> InterpResult<'tcx, Scalar<M::Provenance>>

fn check_wide_ptr_meta( &mut self, meta: MemPlaceMeta<M::Provenance>, pointee: TyAndLayout<'tcx> ) -> InterpResult<'tcx>

fn check_safe_pointer( &mut self, value: &OpTy<'tcx, M::Provenance>, ptr_kind: PointerKind ) -> InterpResult<'tcx>

Check a reference or Box.

fn try_visit_primitive( &mut self, value: &OpTy<'tcx, M::Provenance> ) -> InterpResult<'tcx, bool>

Check if this is a value of primitive type, and if yes check the validity of the value at that type. Return true if the type is indeed primitive.

Note that not all of these have FieldsShape::Primitive, e.g. wide references.

fn visit_scalar( &mut self, scalar: Scalar<M::Provenance>, scalar_layout: ScalarAbi ) -> InterpResult<'tcx>