Trait rustc_const_eval::pattern::PatternFoldable

pub trait PatternFoldable<'tcx>: Sized {
    fn super_fold_with<F: PatternFolder<'tcx>>(&self, folder: &mut F) -> Self;

    fn fold_with<F: PatternFolder<'tcx>>(&self, folder: &mut F) -> Self { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn super_fold_with<F: PatternFolder<'tcx>>(&self, folder: &mut F) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn fold_with<F: PatternFolder<'tcx>>(&self, folder: &mut F) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?