Trait rustc_const_eval::pattern::PatternFolder

pub trait PatternFolder<'tcx>: Sized {
    fn fold_pattern(&mut self, pattern: &Pattern<'tcx>) -> Pattern<'tcx> { ... }

    fn fold_pattern_kind(
        &mut self, 
        kind: &PatternKind<'tcx>
    ) -> PatternKind<'tcx> { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Provided Methods

fn fold_pattern(&mut self, pattern: &Pattern<'tcx>) -> Pattern<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fold_pattern_kind(&mut self, kind: &PatternKind<'tcx>) -> PatternKind<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?