Struct rustc_const_eval::ConstContext

pub struct ConstContext<'a, 'tcx: 'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl<'a, 'tcx> ConstContext<'a, 'tcx>

fn new(tcx: TyCtxt<'a, 'tcx, 'tcx>, body: BodyId) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn with_tables(tcx: TyCtxt<'a, 'tcx, 'tcx>,
tables: &'a TypeckTables<'tcx>)
-> Self

🔬 This is a nightly-only experimental API. (rustc_private)

fn eval(&self, e: &Expr, ty_hint: EvalHint<'tcx>) -> EvalResult

🔬 This is a nightly-only experimental API. (rustc_private)

Evaluate a constant expression in a context where the expression isn't guaranteed to be evaluatable. ty_hint is usually ExprTypeChecked, but a few places need to evaluate constants during type-checking, like computing the length of an array. (See also the FIXME above EvalHint.)

impl<'a, 'tcx> ConstContext<'a, 'tcx>

fn compare_lit_exprs(&self,
span: Span,
a: &Expr,
b: &Expr)
-> Result<Ordering, ErrorReported>

🔬 This is a nightly-only experimental API. (rustc_private)