Struct rustc_const_eval::transform::promote_consts::Validator

struct Validator<'a, 'tcx> {
    ccx: &'a ConstCx<'a, 'tcx>,
    temps: &'a mut IndexVec<Local, TempState>,
}

Checks whether locals that appear in a promotion context (Candidate) are actually promotable.

This wraps an Item, and has access to all fields of that Item via Deref coercion.

Fields
ccx: &'a ConstCx<'a, 'tcx>temps: &'a mut IndexVec<Local, TempState>
Implementations

impl<'tcx> Validator<'_, 'tcx>

fn validate_candidate(
&mut self,
candidate: Candidate
) -> Result<(), Unpromotable>

fn qualif_local<Q: Qualif>(&mut self, local: Local) -> bool

fn validate_local(&mut self, local: Local) -> Result<(), Unpromotable>

fn validate_place(&mut self, place: PlaceRef<'tcx>) -> Result<(), Unpromotable>

fn validate_operand(
&mut self,
operand: &Operand<'tcx>
) -> Result<(), Unpromotable>

fn validate_ref(
&mut self,
kind: BorrowKind,
place: &Place<'tcx>
) -> Result<(), Unpromotable>

fn validate_rvalue(&mut self, rvalue: &Rvalue<'tcx>) -> Result<(), Unpromotable>

fn validate_call(
&mut self,
callee: &Operand<'tcx>,
args: &[Operand<'tcx>]
) -> Result<(), Unpromotable>