Crate rustc_const_eval

Modules

check_consts

Check the bodies of consts, statics and const fns for illegal operations.

const_eval

errors 🔒

interpret

An interpreter for MIR used in CTFE and by miri

util

Statics

CTRL_C_RECEIVED

rustc_driver::main installs a handler that will set this to true if the compiler has been sent a request to shut down, such as by a Ctrl-C. This static lives here because it is only read by the interpreter.

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Traits

ReportErrorExt

Functions

provide