Crate rustc_const_eval

Unstable (rustc_private)

constant evaluation on the HIR and code to validate patterns/matches

Note

This API is completely unstable and subject to change.

Modules

check_match 	 [Unstable]

diagnostics 	 [Unstable]

Structs

ConstEvalErr 	 [Unstable]

Enums

ErrKind 	 [Unstable]

EvalHint 	 [Unstable]
Hint to determine how to evaluate constant expressions which might not be type-checked.

Constants

DIAGNOSTICS 	 [Unstable]

Functions

compare_const_vals 	 [Unstable]

compare_lit_exprs 	 [Unstable]

const_expr_to_pat 	 [Unstable]

eval_const_expr 	 [Unstable]

eval_const_expr_partial 	 [Unstable]
Evaluate a constant expression in a context where the expression isn't guaranteed to be evaluatable. ty_hint is usually ExprTypeChecked, but a few places need to evaluate constants during type-checking, like computing the length of an array. (See also the FIXME above EvalHint.)

eval_repeat_count 	 [Unstable]
Returns the repeat count for a repeating vector expression.

lookup_const_by_id 	 [Unstable]

def_id is the id of the constant.

substs is the monomorphized substitutions for the expression.

lookup_const_fn_by_id 	 [Unstable]

Type Definitions

CastResult 	 [Unstable]

EvalResult 	 [Unstable]

FnArgMap 	 [Unstable]