rustc_const_eval
Module const_eval

Modules

dummy_machine 🔒

error 🔒

eval_queries 🔒

fn_queries 🔒

machine 🔒

valtrees 🔒

Macros

throw_machine_stop_str

Macro for machine-specific InterpError without allocation. (These will never be shown to the user, but they help diagnose ICEs.)

Structs

CompileTimeMachine

Extra machine state for CTFE, and the Machine instance.

DummyMachine

Enums

CheckAlignment

ConstEvalErrKind

The CTFE machine has some custom error kinds.

MemoryKind

ValTreeCreationError 🔒

Constants

VALTREE_MAX_NODES 🔒

Traits

InterpretationResult

Functions

eval_static_initializer_provider

eval_to_allocation_raw_provider

eval_to_const_value_raw_provider

get_span_and_frames

is_parent_const_impl_raw

is_unstable_const_fn

Whether the def_id is an unstable const fn and what feature gate(s) are necessary to enable it.

mk_eval_cx_for_const_val

Create an interpreter context to inspect the given ConstValue. Returns both the context and an OpTy that represents the constant.

provide

tag_for_variant_provider

Computes the tag (if any) for a given type and variant.

try_destructure_mir_constant_for_user_output 🔒

Type Aliases

CompileTimeInterpCx

ValTreeCreationResult 🔒