rustc_const_eval::const_eval
Module eval_queries

Traits

InterpretationResult

Functions

const_validate_mplace 🔒

eval_body_using_ecx 🔒

eval_in_interpreter 🔒

eval_static_initializer_provider

eval_to_allocation_raw_provider

eval_to_const_value_raw_provider

mk_eval_cx_for_const_val

Create an interpreter context to inspect the given ConstValue. Returns both the context and an OpTy that represents the constant.

mk_eval_cx_to_read_const_val 🔒

The InterpCx is only meant to be used to do field and index projections into constants for simd_shuffle and const patterns in match arms.

op_to_const 🔒

This function converts an interpreter value into a MIR constant.

report_eval_error 🔒

report_validation_error 🔒

turn_into_const_value 🔒