Module rustc_const_eval::const_eval::machine

Structs

CompileTimeMachine

Extra machine state for CTFE, and the Machine instance

Enums

CanAccessMutGlobal 🔒

CheckAlignment

MemoryKind

Constants

LINT_TERMINATOR_LIMIT 🔒

When hitting this many interpreted terminators we emit a deny by default lint that notfies the user that their constant takes a long time to evaluate. If that’s what they intended, they can just allow the lint.

PROGRESS_INDICATOR_START 🔒

After this many interpreted terminators, we start emitting progress indicators at every power of two of interpreted terminators.

TINY_LINT_TERMINATOR_LIMIT 🔒

The limit used by -Z tiny-const-eval-limit. This smaller limit is useful for internal tests not needing to run 30s or more to show some behaviour.

Type Aliases

CompileTimeInterpCx 🔒