Module rustc_const_eval::const_eval::valtrees

Functions

branches 🔒

const_to_valtree_inner 🔒

create_valtree_place 🔒

dump_place 🔒

eval_to_valtree 🔒

Evaluates a constant and turns it into a type-level constant value.

reconstruct_place_meta 🔒

Valtrees don’t store the MemPlaceMeta that all dynamically sized values have in the interpreter. This function reconstructs it.

slice_branches 🔒

valtree_into_mplace 🔒

valtree_to_const_value

Converts a ValTree to a ConstValue, which is needed after mir construction has finished.

valtree_to_ref 🔒

Put a valtree into memory and return a reference to that.