Module rustc_const_eval::interpret::eval_context

Structs

InterpCx

Functions

format_interp_error

Turn the given error into a human-readable string. Expects the string to be printed, so if RUSTC_CTFE_BACKTRACE is set this will show a backtrace of the rustc internals that triggered the error.

from_known_layout 🔒

Use the already known layout if given (but sanity check in debug mode), or compute the layout.

mir_assign_valid_types 🔒

Test if it is valid for a MIR assignment to assign src-typed place to dest-typed value. This test should be symmetric, as it is primarily about layout compatibility.