Module rustc_const_eval::interpret::machine

This module contains everything needed to instantiate an interpreter. This separation exists to ensure that no fancy miri features like interpreting common C functions leak into CTFE.

Macros

compile_time_machine

A lot of the flexibility above is just needed for Miri, but all “compile-time” machines (CTFE and ConstProp) use the same instance. Here, we share that code.

Enums

StackPopJump

Data returned by Machine::stack_pop, to provide further control over the popping of the stack frame

Traits

AllocMap

The functionality needed by memory to manage its allocations

Machine

Methods of this trait signifies a point where CTFE evaluation would fail and some use case dependent behaviour can instead be applied.

MayLeak

Whether this kind of memory is allowed to leak