rustc_const_eval::interpret
Module machine

This module contains everything needed to instantiate an interpreter. This separation exists to ensure that no fancy miri features like interpreting common C functions leak into CTFE.

Macros

compile_time_machine

A lot of the flexibility above is just needed for Miri, but all “compile-time” machines (CTFE and ConstProp) use the same instance. Here, we share that code.

Enums

ReturnAction

Data returned by Machine::after_stack_pop, and consumed by InterpCx::return_from_current_stack_frame to determine what actions should be done when returning from a stack frame.

Traits

AllocMap

The functionality needed by memory to manage its allocations

Machine

Methods of this trait signifies a point where CTFE evaluation would fail and some use case dependent behaviour can instead be applied.

MayLeak

Whether this kind of memory is allowed to leak