rustc_const_eval::interpret
Module memory

The memory subsystem.

Generally, we use Pointer to denote memory addresses. However, some operations have a “size”-like parameter, and they take Scalar for the address because if the size is 0, then the pointer can also be a (properly aligned, non-null) integer. It is crucial that these operations call check_align before short-circuiting the empty case!

Structs

AllocInfo

Metadata about an AllocId.

AllocRef

A reference to some allocation that was already bounds-checked for the given region and had the on-access machine hooks run.

AllocRefMut

A reference to some allocation that was already bounds-checked for the given region and had the on-access machine hooks run.

Memory

Enums

AllocKind

The return value of get_alloc_info indicates the “kind” of the allocation.

FnVal

The value of a function pointer.

MemoryKind