Struct rustc_const_eval::ConstContext

pub struct ConstContext<'a, 'tcx: 'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'tcx> ConstContext<'a, 'tcx>

pub fn new(
tcx: TyCtxt<'a, 'tcx, 'tcx>,
param_env_and_substs: ParamEnvAnd<'tcx, &'tcx Substs<'tcx>>,
tables: &'a TypeckTables<'tcx>
) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn eval(&self, e: &'tcx Expr) -> EvalResult<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Evaluate a constant expression in a context where the expression isn't guaranteed to be evaluable.

impl<'a, 'tcx> ConstContext<'a, 'tcx>

pub fn compare_lit_exprs(
&self,
span: Span,
a: &'tcx Expr,
b: &'tcx Expr
) -> Result<Ordering, ErrorReported>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?